Add python interface to mesh projector
* Based on nanobind for Python 3.12+ stable ABI support
* Can be built as dependency-less("static") shared object
* Requires musl-based system for build
* Added Alpine Dockerfile for py4dproject build
* Added preset for configuring with py4dproject building
* Adjusted pipeline to generate alpine images as well
* Adjusted code for CMake 4.x.y compatbility
* Fixed warnings when building without LTO

+++ src/stmesh/vtk_writer.cpp
@@ -5,6 +5,7 @@
 #include <bit>
 #include <cmath>
 #include <cstddef>
+#include <cstdint>
 #include <filesystem>
 #include <fstream>
 #include <random>
@@ -168,7 +169,8 @@ void writeVTUFile(const std::filesystem::path &directory, const std::string_view
       const vtkIdType id = grid->InsertNextCell(VTK_POLYHEDRON, static_cast<vtkIdType>(polyhedron.point_ids_.size()),
                                                 polyhedron.point_ids_.data(), num_facets, faces->GetPointer(0));
       id_array->InsertTuple1(id, static_cast<double>(polyhedron.polyhedron_id_));
-      random_array->InsertTuple1(id, static_cast<double>(std::mt19937(polyhedron.polyhedron_id_)()));
+      random_array->InsertTuple1(
+          id, static_cast<double>(std::mt19937(static_cast<uint_fast32_t>(polyhedron.polyhedron_id_))()));
     }
     grid->GetCellData()->AddArray(id_array);
     grid->GetCellData()->AddArray(random_array);
